Updating routes to add additional user protection verifying user accessing route to be user in token

+++ model/postsModel.js
@@ -5,6 +5,10 @@ class Posts{
         const [rows] = await dbQuery('INSERT INTO posts (post_content, group_id, user_id, incognito, post_user_quit) VALUES (?, ?, ?, ?, ?)', [postContent, groupId, userId, incognito, postUserQuit]);
         return rows;
     }
+    static async getPost(postId){
+        const [rows] = await dbQuery('SELECT post_created FROM posts WHERE post_id = ? ', [postId]);
+        return rows;
+    }
     static async selectAllPostWithUser(groupId){
         const [rows] = await dbQuery('SELECT * FROM  posts JOIN  users ON posts.user_id = users.user_id WHERE group_id = ? ORDER BY post_created DESC', [groupId]);
         return rows;
@@ -17,12 +21,12 @@ class Posts{
         const [rows] = await dbQuery('SELECT * FROM  comments JOIN  users ON comments.user_id = users.user_id WHERE post_id = ?', [postId]);
         return rows;
     }
-    static  async deletePosts(postId){
-        const [rows] = await dbQuery('DELETE FROM posts WHERE post_id = ?',[postId]);
+    static  async deletePosts(postId, userId){
+        const [rows] = await dbQuery('DELETE FROM posts WHERE post_id = ? AND user_id = ?',[postId, userId]);
         return rows;
     }
-    static async updateUserPost(postId, postContent){
-        const [rows] = await dbQuery('UPDATE posts SET post_content = ? WHERE post_id = ?',[postContent, postId]);
+    static async updateUserPost(postId, postContent, userId){
+        const [rows] = await dbQuery('UPDATE posts SET post_content = ? WHERE post_id = ? AND user_id = ?',[postContent, postId, userId]);
         return rows;
     }
     static async updatePostStatusOnceUserQuitsGroup(postUserQuit , groupId, userId){
